Make change to login happen faster, add header during app load
