+++ src/fetch.js
@@ -2,8 +2,7 @@ import * as isomorphicFetch from 'isomorphic-fetch';
 import _ from 'lodash';
 
 import { store } from '~/store';
-import { checkLogin } from '~/session';
-import { setToken } from '~/actions/authentication';
+import * as session from '~/session';
 import { API_ROOT } from './constants';
 
 /*
@@ -12,12 +11,6 @@ import { API_ROOT } from './constants';
  */
 export function rawFetch(...args) {
   return isomorphicFetch['default'](...args);
-}
-
-function expireSession() {
-  const next = { location: window.location };
-  store.dispatch(setToken(null, null, null, null));
-  checkLogin(next);
 }
 
 export function fetch(token, _path, _options) {
@@ -54,7 +47,7 @@ export function fetch(token, _path, _options) {
       const { status } = response;
       if (status >= 400) {
         if (status === 401) {
-          expireSession();
+          dispatch(session.expire);
         }
 
         reject(response);
